[expo-cli] Allow manully provided credentials when keytool is missing
